float to int

+++ Blur.cpp
@@ -1,6 +1,6 @@
 #include "Blur.h"
 
-const float Blur::_weights[Convolutional::wSize()][Convolutional::wSize()] = { {1, 1, 1}, {1, 1, 1}, {1, 1, 1} };
+const int Blur::_weights[Convolutional::wSize()][Convolutional::wSize()] = { {1, 1, 1}, {1, 1, 1}, {1, 1, 1} };
 
 Blur::Blur(int up, int left, int bottom, int right) : Convolutional(up, left, bottom, right) {
 }
